+++ app/src/main/java/com/example/playem/btmanager/blehandlers/DISPnpIDReadRequest.java
@@ -7,7 +7,7 @@ import android.bluetooth.BluetoothGatt;
 import android.bluetooth.BluetoothGattCharacteristic;
 import android.bluetooth.BluetoothGattServer;
 
-import com.example.playem.btmanager.GattResponse;
+import com.example.playem.btmanager.blehandlers.interfaces.GattResponse;
 import com.example.playem.btmanager.blehandlers.interfaces.BLECharacteristicsReadRequest;
 
 public class DISPnpIDReadRequest implements BLECharacteristicsReadRequest {
